MULE-10866: support for descriptors in mule-api to be later consume in the deployments modules
_ Moving MulePluginModelBuilder as an inner class of MulePluginModel
