+++ src/main/java/org/mule/runtime/api/deployment/meta/MulePluginModel.java
@@ -6,9 +6,15 @@
  */
 package org.mule.runtime.api.deployment.meta;
 
+import static java.util.Optional.empty;
+import static java.util.Optional.of;
 import static java.util.Optional.ofNullable;
+import static org.apache.commons.lang3.StringUtils.isBlank;
+import static org.mule.runtime.api.util.Preconditions.checkArgument;
 import org.mule.runtime.api.meta.model.ExtensionModel;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Optional;
 
 /**
@@ -23,22 +29,14 @@ import java.util.Optional;
  */
 public class MulePluginModel {
 
-  final private String name;
-  final private String minMuleVersion;
-  final private MulePluginLoaderDescriptor classLoaderModelLoaderDescriptor;
-  final private MulePluginLoaderDescriptor extensionModelLoaderDescriptor;
+  private final String name;
+  private final String minMuleVersion;
+  private final MulePluginLoaderDescriptor extensionModelLoaderDescriptor;
+  private final MulePluginLoaderDescriptor classLoaderModelLoaderDescriptor;
 
-  /**
-   * Creates an immutable implementation of {@link MulePluginModel}
-   *
-   * @param name of the artifact
-   * @param minMuleVersion minimal Mule version it runs on
-   * @param classLoaderModelLoaderDescriptor information related to the plugin to generate an {@link ClassLoader}
-   * @param extensionModelLoaderDescriptor information related to the plugin to generate an {@link ExtensionModel}
-   */
-  public MulePluginModel(String name, String minMuleVersion,
-                         MulePluginLoaderDescriptor classLoaderModelLoaderDescriptor,
-                         MulePluginLoaderDescriptor extensionModelLoaderDescriptor) {
+  private MulePluginModel(String name, String minMuleVersion,
+                          MulePluginLoaderDescriptor classLoaderModelLoaderDescriptor,
+                          MulePluginLoaderDescriptor extensionModelLoaderDescriptor) {
     this.name = name;
     this.minMuleVersion = minMuleVersion;
     this.extensionModelLoaderDescriptor = extensionModelLoaderDescriptor;
@@ -59,5 +57,125 @@ public class MulePluginModel {
 
   public Optional<MulePluginLoaderDescriptor> getExtensionModelLoaderDescriptor() {
     return ofNullable(extensionModelLoaderDescriptor);
+  }
+
+  /**
+   * A builder to create instances of {@link MulePluginModel}.
+   *
+   * @since 1.0
+   */
+  public static class MulePluginModelBuilder {
+
+    private String name;
+    private String minMuleVersion;
+    private Optional<MulePluginPropertyBuilder> classLoaderDescriptorBuilder = empty();
+    private Optional<MulePluginPropertyBuilder> extensionModelDescriptorBuilder = empty();
+
+    /**
+     * Sets the describer's name
+     *
+     * @param name the name the describer will have
+     * @return {@code this} builder
+     */
+    public MulePluginModelBuilder setName(String name) {
+      this.name = name;
+      return this;
+    }
+
+    /**
+     * Sets the describer's minimum Mule Runtime version that requires to work correctly
+     *
+     * @param muleVersion of the describer
+     * @return {@code this} builder
+     */
+    public MulePluginModelBuilder setMinMuleVersion(String muleVersion) {
+      this.minMuleVersion = muleVersion;
+      return this;
+    }
+
+    /**
+     * @return a {@link MulePluginPropertyBuilder} to populate the {@link ClassLoader} describer with the ID and
+     * any additional attributes
+     */
+    public MulePluginPropertyBuilder withClassLoaderModelDescriber() {
+      if (!classLoaderDescriptorBuilder.isPresent()) {
+        classLoaderDescriptorBuilder = of(new MulePluginPropertyBuilder());
+      }
+      return classLoaderDescriptorBuilder.get();
+    }
+
+    /**
+     * @return a {@link MulePluginPropertyBuilder} to populate the {@link ExtensionModel} describer with the ID and
+     * any additional attributes
+     */
+    public MulePluginPropertyBuilder withExtensionModelDescriber() {
+      if (!extensionModelDescriptorBuilder.isPresent()) {
+        extensionModelDescriptorBuilder = of(new MulePluginPropertyBuilder());
+      }
+      return extensionModelDescriptorBuilder.get();
+    }
+
+    /**
+     * @return a well formed {@link MulePluginModel}
+     */
+    public MulePluginModel build() {
+      checkArgument(!isBlank(name), "name cannot be a blank");
+      checkArgument(minMuleVersion != null, "minMuleVersion cannot be null");
+      return new MulePluginModel(name, minMuleVersion,
+                                 classLoaderDescriptorBuilder.isPresent() ? classLoaderDescriptorBuilder.get().build() : null,
+                                 extensionModelDescriptorBuilder.isPresent() ? extensionModelDescriptorBuilder.get().build()
+                                     : null);
+    }
+
+    /**
+     * A builder to create instances of {@link MulePluginLoaderDescriptor}.
+     * <p>
+     * A new instance of this class should be used per each manifest to be
+     * created. The created instances will be immutable.
+     *
+     * @since 4.0
+     */
+    public static final class MulePluginPropertyBuilder {
+
+      private String id;
+      private final Map<String, Object> properties = new HashMap<>();
+
+      /**
+       * Sets the describer's ID
+       *
+       * @param id the ID to be set
+       * @return {@code this} builder
+       */
+      public MulePluginPropertyBuilder setId(String id) {
+        this.id = id;
+        return this;
+      }
+
+      /**
+       * Sets the given property on the describer.
+       * <p>
+       * If a value is already associated with the {@code key}, then
+       * it is overridden with the new {@code value}
+       *
+       * @param key   the property's key
+       * @param value the property's value
+       * @return {@code this} builder
+       */
+      public MulePluginPropertyBuilder addProperty(String key, Object value) {
+        checkArgument(!isBlank(key), "key cannot be a blank key");
+        properties.put(key, value);
+        return this;
+      }
+
+      /**
+       * Creates and returns a new {@link MulePluginLoaderDescriptor} according to the values set
+       *
+       * @return a {@link MulePluginLoaderDescriptor}
+       */
+      private MulePluginLoaderDescriptor build() {
+        checkArgument(!isBlank(id), "ID cannot be a blank");
+        return new MulePluginLoaderDescriptor(id, properties);
+      }
+    }
   }
 }
